Optimize trapz.* types

+++ src/trapz/trapz.h
@@ -2,7 +2,8 @@
 #define _TRAPZ_H_
 
 #include <math.h>
+#include <stdint.h>
 
-double catools_trapz(int n, double *x, double *y);
+double catools_trapz(uint16_t n, double *x, double *y);
 
 #endif // _TRAPZ_H_
